requirements.txt
pytorch-lightning==1.7.5
torch==1.12.0
torchvision==0.13.0+cu113
numpy==1.22.4
einops==0.6.0
pandas==1.4.2
Pillow==9.0.1
rasterio==1.3.0
matplotlib==3.5.2